api/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
boto3==1.34.162
cryptography==43.0.1
gunicorn==23.0.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pdfplumber==0.11
python-docx==0.8.11
openpyxl==3.1
pandas==2.2
PyJWT>=2.8.0
python-pptx==0.6.23
Flask-Migrate
